django-housing/requirements.txt
Django==5.1.3
psycopg2-binary==2.9.10
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0 
channels==4.0.0
daphne==4.0.0
drf-yasg==1.21.7